Add support for Enlighten's directional lightmaps.
It store 3 floats for color, 3 for lighting main direction information,
and 1 float to ensure that the directional term evaluates to 1 when the
surface normal aligns with normal used when baking.

NOTE: A directional map can be encoded per RGB channel which puts the
memory cost at the cost of L1 SH but at a worse quality. This
implementation with a single directional map is provided as a cheap
alternative at the expense of quality.

Reference: https://static.docs.arm.com/100837/0308/enlighten_3-08_sdk_documentation__100837_0308_00_en.pdf